OpenART/pyproject.toml
[project]
name = "openart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "rembg>=2.0.67",
    "tensorflow-cpu>=2.18.0",
]
